pyproject.toml
[tool.poetry]
name = "transaction-analysis"
version = "0.1.0"
description = ""
authors = ["Natalia Pochekutova <natalipochekutova51@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"
openpyxl = "^3.1.5"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.0"
black = "^24.8.0"
isort = "^5.13.2"
mypy = "^1.11.1"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
exclude = """.git"""

[tool.isort]
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
check_untyped_defs = true
exclude = 'venv'

[tool.pytest.ini_options]
addopts = "--cov --cov-report term-missing"